all the stuff for a LocationListener
There are (at least) a trio of horrible hacks here

1) until rubicon-java supports passing arrays of String, we can't call
requestPermissions from Python. So we have a kotlin kelper in
Permissions.kt to do it for us, and we have the Makefile copy that
file into the generated android sources

2) the java "thing" (technical term) created by rubicon-java's
JavaInterface does not implement hashCode, which is a problem as
Android's location manager code wants to store LocationListeners in a
hash. We add it to self._methods in __init__ when we create our
LocListener object

3) we need location permissions added to AndroidManifest.xml, which
is generated by briefcase, so we have the makefile edit that file.

There are unquestionably better ways of doing each of these
things. These are workarounds until I have the understanding of what
those better ways are